requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-flask>=1.3.0
